pdf-parser/requirements.txt
PyMuPDF>=1.23.0
Markdown>=3.5.0
Pillow>=10.0.0
anthropic>=0.18.0
tenacity>=8.2.0
python-dotenv>=1.0.0

